requirements.txt
selenium
mysql-connector-python
beautifulsoup4
requests